service.c: cancel online check when removing service
When the network is removed e.g unplugging the wire,
network.c call connman_service_remove_from_network,
until the refcount is positive the service remains
in the list and we have to wait the online timeout
to release the last ref on the service.

Better to cancel the online check in case of disconnection
to remove quickly the service.